Assignment 5
Assignment 5 code. The initial assignment was to get familiar with using
Docker to host Postgres and insert some records using a provided Go
example. The second part of the assignment is to then use Go to create
an MSDS Course catalog, similar to what we have done previously, and
then save it to the database. In other words, modify the code provdied
to store different types of data.